+++ FinalBases/src/main/java/org/example/ManejadorDepartamentos.java
@@ -10,10 +10,7 @@ import org.bson.conversions.Bson;
 
 //POSTGRESQL
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
+import java.sql.*;
 
 import javax.swing.*;
 import java.util.ArrayList;
@@ -112,6 +109,36 @@ public class ManejadorDepartamentos {
             JOptionPane.showMessageDialog(null, "No se encontraron departamentos con el código especificado.");
         }
 
+
+
+        //POSTGRESQL
+        String url = "jdbc:postgresql://localhost:5432/Eventos";
+        String usuario = "postgres";
+        String contraseña = "admin123";
+
+        try {
+            Connection conexión = DriverManager.getConnection(url, usuario, contraseña);
+
+            // Consulta SQL para obtener los departamentos
+            String sql = "SELECT COD_DEPARTAMENTO, NOM_DEPARTAMENTO FROM DEPARTAMENTOS";
+            PreparedStatement declaración = conexión.prepareStatement(sql);
+
+            ResultSet resultado = declaración.executeQuery();
+
+            // Recorre los resultados e imprime los datos de los departamentos
+            while (resultado.next()) {
+                int codDepartamento = resultado.getInt("COD_DEPARTAMENTO");
+                String nomDepartamento = resultado.getString("NOM_DEPARTAMENTO");
+                System.out.println("POSTGRESQL");
+                System.out.println("Código: " + codDepartamento + ", Nombre: " + nomDepartamento);
+            }
+
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
+
     }
 
     public static void modificarDepartamento() {
@@ -149,6 +176,36 @@ public class ManejadorDepartamentos {
             JOptionPane.showMessageDialog(null, "No se encontró un departamento con el código especificado.");
         }
 
+
+        //POSTGRESQL
+        String url = "jdbc:postgresql://localhost:5432/Eventos";
+        String usuario = "postgres";
+        String contraseña = "admin123";
+
+        try {
+            Connection conexión = DriverManager.getConnection(url, usuario, contraseña);
+
+            // Consulta SQL de actualización de departamento
+            String sql = "UPDATE DEPARTAMENTOS SET NOM_DEPARTAMENTO = ? WHERE COD_DEPARTAMENTO = ?";
+            PreparedStatement declaración = conexión.prepareStatement(sql);
+
+            // Valores para la actualización
+            declaración.setString(1, nuevoNombre); // Nuevo nombre del departamento
+            declaración.setInt(2, codigoModificar); // Código del departamento a actualizar
+
+            int filasActualizadas = declaración.executeUpdate();
+            if (filasActualizadas > 0) {
+                System.out.println("Departamento actualizado correctamente.");
+            } else {
+                System.out.println("No se encontró el departamento a actualizar.");
+            }
+
+            declaración.close();
+            conexión.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
     }
 
     public static void eliminarDepartamento() {
@@ -167,17 +224,49 @@ public class ManejadorDepartamentos {
         // Se pide el código de el departamento a eliminar
         int codigoEliminar = Integer.parseInt(JOptionPane.showInputDialog(null, "Por favor ingrese el código del departamento a eliminar"));
 
-        // Crear un filtro de eliminación para buscar por el código de la departamento
-        Bson filtro = Filters.eq("cod_departamento", codigoEliminar);
 
-        // Eliminar el departamento utilizando el filtro
-        DeleteResult resultado = departamentosCollection.deleteOne(filtro);
 
-        // Verificar si se eliminó correctamente el departamento
-        if (resultado.getDeletedCount() > 0) {
-            JOptionPane.showMessageDialog(null, "El departamento se eliminó correctamente.");
-        } else {
-            JOptionPane.showMessageDialog(null, "No se encontró una departamento con el código especificado.");
+        //POSTGRESQL
+        String url = "jdbc:postgresql://localhost:5432/Eventos";
+        String usuario = "postgres";
+        String contraseña = "admin123";
+
+        try {
+            Connection conexión = DriverManager.getConnection(url, usuario, contraseña);
+
+            // Consulta SQL de eliminación de departamento
+            String sql = "DELETE FROM DEPARTAMENTOS WHERE COD_DEPARTAMENTO = ?";
+            PreparedStatement declaración = conexión.prepareStatement(sql);
+
+            // Valor para la eliminación
+            declaración.setInt(1, codigoEliminar); // Código del departamento a eliminar
+
+            int filasEliminadas = declaración.executeUpdate();
+            if (filasEliminadas > 0) {
+                JOptionPane.showMessageDialog(null, "El departamento se eliminó correctamente de postgres.");
+
+                //Borrado mongo
+
+                // Crear un filtro de eliminación para buscar por el código de la departamento
+                Bson filtro = Filters.eq("cod_departamento", codigoEliminar);
+
+                // Eliminar el departamento utilizando el filtro
+                DeleteResult resultado = departamentosCollection.deleteOne(filtro);
+
+                // Verificar si se eliminó correctamente el departamento
+                if (resultado.getDeletedCount() > 0) {
+                    JOptionPane.showMessageDialog(null, "El departamento se eliminó correctamente de mongo.");
+                } else {
+                    JOptionPane.showMessageDialog(null, "No se encontró una departamento con el código especificado.");
+                }
+            } else {
+                System.out.println("No se encontró el departamento a eliminar.");
+            }
+
+            declaración.close();
+            conexión.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
         }
 
     }
